+++ src/components/ui/icon/CopyAddressIcon.tsx
@@ -1,0 +1,20 @@
+const CopyAddressIcon = () => {
+  return (
+    <svg width="24" height="25" viewBox="0 0 24 25" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M16.5008 5.89082H20.7195C20.8742 5.89082 21.0008 5.76426 21.0008 5.60957C21.0008 5.2252 20.8273 4.85957 20.532 4.61113L16.918 1.60176C16.6883 1.40957 16.2523 1.25488 15.9523 1.25488C15.7602 1.25488 15.6055 1.40957 15.6055 1.60176V5.0002C15.6102 5.49238 16.0086 5.89082 16.5008 5.89082Z"
+        fill="#B3916A"
+      />
+      <path
+        d="M14.3906 5V1.25H8.25C7.425 1.25 6.75 1.925 6.75 2.75V18.5C6.75 19.325 7.425 20 8.25 20H19.5C20.325 20 21 19.325 21 18.5V7.10938H16.5C15.3375 7.10938 14.3906 6.1625 14.3906 5Z"
+        fill="#B3916A"
+      />
+      <path
+        d="M5.4375 19.8125V4.25H4.5C3.675 4.25 3 4.925 3 5.75V22.25C3 23.075 3.675 23.75 4.5 23.75H16.5C17.325 23.75 18 23.075 18 22.25V21.3125H6.9375C6.1125 21.3125 5.4375 20.6375 5.4375 19.8125Z"
+        fill="#B3916A"
+      />
+    </svg>
+  );
+};
+
+export default CopyAddressIcon;
