+++ src/components/ui/icon/RunningIcon.tsx
@@ -1,0 +1,12 @@
+const RunningIcon = () => {
+  return (
+    <svg width="17" height="22" viewBox="0 0 17 22" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M11.4901 4.48047C12.5901 4.48047 13.4901 3.58047 13.4901 2.48047C13.4901 1.38047 12.5901 0.480469 11.4901 0.480469C10.3901 0.480469 9.49014 1.38047 9.49014 2.48047C9.49014 3.58047 10.3901 4.48047 11.4901 4.48047ZM7.89014 18.3805L8.89014 13.9805L10.9901 15.9805V21.9805H12.9901V14.4805L10.8901 12.4805L11.4901 9.48047C12.7901 10.9805 14.7901 11.9805 16.9901 11.9805V9.98047C15.0901 9.98047 13.4901 8.98047 12.6901 7.58047L11.6901 5.98047C11.2901 5.38047 10.6901 4.98047 9.99014 4.98047C9.69014 4.98047 9.49014 5.08047 9.19014 5.08047L3.99014 7.28047V11.9805H5.99014V8.58047L7.79014 7.88047L6.19014 15.9805L1.29014 14.9805L0.890137 16.9805L7.89014 18.3805Z"
+        fill="#A0A0A0"
+      />
+    </svg>
+  );
+};
+
+export default RunningIcon;
