+++ src/components/ui/icon/SubwayIcon.tsx
@@ -1,0 +1,11 @@
+const SubwayIcon = () => {
+  return (
+    <svg width="18" height="20" viewBox="0 0 18 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path
+        d="M9 0.25C4.60013 0.25 0.75 0.750062 0.75 4.25003V14.25C0.75 16.2002 2.29978 17.7499 4.25002 17.7499L3 19V19.75H15V19L13.75 17.7499C15.7002 17.7499 17.25 16.2002 17.25 14.25V4.25003C17.25 0.750062 13.3999 0.25 9 0.25ZM4.5 16C3.64987 16 3 15.3501 3 14.5C3 13.6498 3.64987 13 4.5 13C5.35012 13 6 13.6498 6 14.5C6 15.3501 5.35012 16 4.5 16ZM8.25 9.25H3V4.75H8.25V9.25ZM13.5 16C12.6499 16 12 15.3501 12 14.5C12 13.6498 12.6499 13 13.5 13C14.3501 13 15 13.6498 15 14.5C15 15.3501 14.3501 16 13.5 16ZM15 9.25H9.75V4.75H15V9.25Z"
+        fill="#A0A0A0"
+      />
+    </svg>
+  );
+};
+export default SubwayIcon;
